Fix CVE scan false positive (#629)
`cve-bin-tool` uses the `product`, `vendor` and `version` of a language
package (e.g. `JMESpath` in this case) to check if there is a
corresponding vulnerability within its database. The issue with this is
that `JMESpath` has multiple packages for different languages, where
their versions are not aligned and there is no differentiation between
them in the CVE database. This means that the critical CVE which affects
the Ruby version of `JMESpath` for versions `< 1.6.1` is flagged for the
latest, unaffected Python package version `1.0.1`. The "solution" here
is to triage that particular package so that it does not appear in the
output of the CVE scanner.