feat: Implement duplicate `date.py` functions from Memgraph
GraphQL support needs `date.convert_format`, currently implemented in
Memgraph's `query_modules`. When we build the Memgraph+MAGE image,
MAGE's `date.py` overwrites Memgraph's. As there is no way to separate
filename from the function's namespace, the best solution (for now) is
to duplicate the code and tests here. When we combine both in a monorepo, the
duplicated code can be removed.